Revamped all examples. Closes #65
