+++ examples/src/tiledb_dense_read_global.cc
@@ -1,10 +1,11 @@
 /**
- * @file   tiledb_read_sparse_1.cc
+ * @file   tiledb_dense_read_global.cc
  *
  * @section LICENSE
  *
  * The MIT License
  *
+ * @copyright Copyright (c) 2017 TileDB, Inc.
  * @copyright Copyright (c) 2016 MIT and Intel Corporation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,38 +28,41 @@
  *
  * @section DESCRIPTION
  *
- * It shows how to read a complete sparse array.
+ * It shows how to read a complete dense array in the global cell order.
+ *
+ * You need to run the following to make it work:
+ *
+ * $ ./tiledb_dense_create
+ * $ ./tiledb_dense_write_global_1
+ * $ ./tiledb_dense_read_global
  */
 
-#include "tiledb.h"
+#include <tiledb.h>
 
-#include <iostream>
+#include <cstdio>
 
 int main(int argc, char** argv) {
-  // Initialize context with the default configuration parameters
+  // Create TileDB context
   tiledb_ctx_t* ctx;
   tiledb_ctx_create(&ctx);
 
   // Prepare cell buffers
-  int buffer_a1[10];
-  uint64_t buffer_a2[10];
-  char buffer_var_a2[30];
-  float buffer_a3[20];
-  int64_t buffer_coords[20];
-  void* buffers[] = {
-      buffer_a1, buffer_a2, buffer_var_a2, buffer_a3, buffer_coords};
+  int buffer_a1[16];
+  uint64_t buffer_a2[16];
+  char buffer_var_a2[40];
+  float buffer_a3[32];
+  void* buffers[] = {buffer_a1, buffer_a2, buffer_var_a2, buffer_a3};
   uint64_t buffer_sizes[] = {sizeof(buffer_a1),
                              sizeof(buffer_a2),
                              sizeof(buffer_var_a2),
-                             sizeof(buffer_a3),
-                             sizeof(buffer_coords)};
+                             sizeof(buffer_a3)};
 
   // Create query
   tiledb_query_t* query;
   tiledb_query_create(
       ctx,
       &query,
-      "my_sparse_array",
+      "my_dense_array",
       TILEDB_READ,
       TILEDB_GLOBAL_ORDER,
       nullptr,
@@ -71,14 +75,15 @@ int main(int argc, char** argv) {
   tiledb_query_submit(ctx, query);
 
   // Print cell values
-  int64_t result_num = buffer_sizes[0] / sizeof(int);
-  printf("coords\t a1\t   a2\t     (a3.first, a3.second)\n");
-  printf("--------------------------------------------------\n");
-  for (int i = 0; i < result_num; ++i) {
-    printf("(%lld, %lld)", buffer_coords[2 * i], buffer_coords[2 * i + 1]);
-    printf("\t %3d", buffer_a1[i]);
-    size_t var_size = (i != result_num - 1) ? buffer_a2[i + 1] - buffer_a2[i] :
-                                              buffer_sizes[2] - buffer_a2[i];
+  uint64_t result_num = buffer_sizes[0] / sizeof(int);
+  printf("result num: %llu\n\n", result_num);
+  printf(" a1\t    a2\t   (a3.first, a3.second)\n");
+  printf("-----------------------------------------\n");
+  for (uint64_t i = 0; i < result_num; ++i) {
+    printf("%3d", buffer_a1[i]);
+    uint64_t var_size = (i != result_num - 1) ?
+                            buffer_a2[i + 1] - buffer_a2[i] :
+                            buffer_sizes[2] - buffer_a2[i];
     printf("\t %4.*s", int(var_size), &buffer_var_a2[buffer_a2[i]]);
     printf("\t\t (%5.1f, %5.1f)\n", buffer_a3[2 * i], buffer_a3[2 * i + 1]);
   }
