+++ examples/src/tiledb_dense_write_global_1.cc
@@ -1,10 +1,11 @@
 /**
- * @file   tiledb_write_dense_sorted.cc
+ * @file   tiledb_dense_write_global_1.cc
  *
  * @section LICENSE
  *
  * The MIT License
  *
+ * @copyright Copyright (c) 2017 TileDB, Inc.
  * @copyright Copyright (c) 2016 MIT and Intel Corporation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,48 +28,91 @@
  *
  * @section DESCRIPTION
  *
- * It shows how to write to a dense array, providing the array cells sorted
- * in row-major order within the specified subarray. TileDB will properly
- * re-organize the cells into the global cell order, prior to writing them
- * on the disk.
+ * It shows how to write an entire dense array in a single write.
+ *
+ * You need to run the following to make this work:
+ *
+ * ./tiledb_dense_create
+ * ./tiledb_dense_write_global_1
  */
 
-#include "tiledb.h"
+#include <tiledb.h>
 
 int main() {
   // Initialize context with the default configuration parameters
   tiledb_ctx_t* ctx;
   tiledb_ctx_create(&ctx);
 
-  // Set the subarray where the write will focus on
-  int64_t subarray[] = {2, 3, 2, 4};
-
   // Prepare cell buffers
-  int buffer_a1[] = {9, 12, 13, 11, 14, 15};
-  uint64_t buffer_a2[] = {0, 2, 3, 5, 9, 12};
-  char buffer_var_a2[] = "jjmnnllllooopppp";
+  // clang-format on
+  int buffer_a1[] = {
+      0,
+      1,
+      2,
+      3,  // Upper left tile
+      4,
+      5,
+      6,
+      7,  // Upper right tile
+      8,
+      9,
+      10,
+      11,  // Lower left tile
+      12,
+      13,
+      14,
+      15  // Lower right tile
+  };
+  uint64_t buffer_a2[] = {
+      0,
+      1,
+      3,
+      6,  // Upper left tile
+      10,
+      11,
+      13,
+      16,  // Upper right tile
+      20,
+      21,
+      23,
+      26,  // Lower left tile
+      30,
+      31,
+      33,
+      36  // Lower right tile
+  };
+  char buffer_var_a2[] =
+      "abbcccdddd"   // Upper left tile
+      "effggghhhh"   // Upper right tile
+      "ijjkkkllll"   // Lower left tile
+      "mnnooopppp";  // Lower right tile
   float buffer_a3[] = {
-      9.1, 9.2, 12.1, 12.2, 13.1, 13.2, 11.1, 11.2, 14.1, 14.2, 15.1, 15.2};
+      0.1,  0.2,  1.1,  1.2,  2.1,  2.2,  3.1,  3.2,   // Upper left tile
+      4.1,  4.2,  5.1,  5.2,  6.1,  6.2,  7.1,  7.2,   // Upper right tile
+      8.1,  8.2,  9.1,  9.2,  10.1, 10.2, 11.1, 11.2,  // Lower left tile
+      12.1, 12.2, 13.1, 13.2, 14.1, 14.2, 15.1, 15.2,  // Lower right tile
+  };
   void* buffers[] = {buffer_a1, buffer_a2, buffer_var_a2, buffer_a3};
   uint64_t buffer_sizes[] = {
       sizeof(buffer_a1),
       sizeof(buffer_a2),
       sizeof(buffer_var_a2) - 1,  // No need to store the last '\0' character
       sizeof(buffer_a3)};
+  // clang-format off
 
   // Create query
   tiledb_query_t* query;
   tiledb_query_create(
-      ctx,
-      &query,
-      "my_dense_array",
-      TILEDB_WRITE,
-      TILEDB_COL_MAJOR,
-      subarray,
-      nullptr,
-      0,
-      buffers,
-      buffer_sizes);
+    ctx,
+    &query,
+    "my_dense_array",
+    TILEDB_WRITE,
+    TILEDB_GLOBAL_ORDER,
+    nullptr,
+    nullptr,
+    0,
+    buffers,
+    buffer_sizes);
 
   // Submit query
   tiledb_query_submit(ctx, query);
