+++ examples/src/tiledb_sparse_write_unordered_1.cc
@@ -1,10 +1,11 @@
 /**
- * @file   tiledb_read_dense_sorted.cc
+ * @file   tiledb_sparse_write_unordered_1.cc
  *
  * @section LICENSE
  *
  * The MIT License
  *
+ * @copyright Copyright (c) 2017 TileDB, Inc.
  * @copyright Copyright (c) 2016 MIT and Intel Corporation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,58 +28,60 @@
  *
  * @section DESCRIPTION
  *
- * It shows how to read from a dense array, constraining the read
- * to a specific subarray and subset of attributes. The cells are copied to the
- * input buffers sorted in row-major order within the selected subarray.
+ * It shows how to write unordered cells to a sparse array in a single write.
+ *
+ * You need to run the following to make this work:
+ *
+ * ./tiledb_sparse_create
+ * ./tiledb_sparse_write_unordered_1
  */
 
-#include <cstdio>
-#include "tiledb.h"
+#include <tiledb.h>
 
 int main() {
-  // Initialize context with the default configuration parameters
+  // Create TileDB context
   tiledb_ctx_t* ctx;
   tiledb_ctx_create(&ctx);
 
-  // Subarray and attributes
-  int64_t subarray[] = {3, 4, 2, 4};
-  const char* attributes[] = {"a1"};
-
   // Prepare cell buffers
-  int buffer_a1[3];
-  void* buffers[] = {buffer_a1};
-  uint64_t buffer_sizes[] = {sizeof(buffer_a1)};
+  // clang-format off
+  int buffer_a1[] = { 7, 5, 0, 6, 4, 3, 1, 2 };
+  uint64_t buffer_a2[] = { 0, 4, 6, 7, 10, 11, 15, 17 };
+  char buffer_var_a2[] = "hhhhffagggeddddbbccc";
+  float buffer_a3[] =
+  {
+      7.1,  7.2,  5.1,  5.2,  0.1,  0.2,  6.1,  6.2,
+      4.1,  4.2,  3.1,  3.2,  1.1,  1.2,  2.1,  2.2
+  };
+  uint64_t buffer_coords[] = { 3, 4, 4, 2, 1, 1, 3, 3, 3, 1, 2, 3, 1, 2, 1, 4 };
+  void* buffers[] =
+      { buffer_a1, buffer_a2, buffer_var_a2, buffer_a3, buffer_coords };
+  uint64_t buffer_sizes[] =
+  {
+      sizeof(buffer_a1),
+      sizeof(buffer_a2),
+      sizeof(buffer_var_a2)-1,  // No need to store the last '\0' character
+      sizeof(buffer_a3),
+      sizeof(buffer_coords)
+  };
+  // clang-format on
 
   // Create query
   tiledb_query_t* query;
   tiledb_query_create(
       ctx,
       &query,
-      "my_dense_array",
-      TILEDB_READ,
-      TILEDB_ROW_MAJOR,
-      subarray,
-      attributes,
-      1,
+      "my_sparse_array",
+      TILEDB_WRITE,
+      TILEDB_UNORDERED,
+      nullptr,
+      nullptr,
+      0,
       buffers,
       buffer_sizes);
 
-  // Loop until no overflow
-  printf(" a1\n----\n");
-  tiledb_query_status_t status;
-  do {
-    // Read from array_schema
-    printf("Reading cells...\n");
-    tiledb_query_submit(ctx, query);
-
-    // Print cell values
-    int64_t result_num = buffer_sizes[0] / sizeof(int);
-    for (int i = 0; i < result_num; ++i)
-      printf("%3d\n", buffer_a1[i]);
-
-    // Get overflow
-    tiledb_query_get_attribute_status(ctx, query, "a1", &status);
-  } while (status == TILEDB_INCOMPLETE);
+  // Submit query
+  tiledb_query_submit(ctx, query);
 
   // Clean up
   tiledb_query_free(ctx, query);
